+++ frontend/src/main/java/org/abs_models/backend/java/observing/TaskView.java
@@ -37,7 +37,7 @@ public interface TaskView {
 
     void registerTaskListener(TaskObserver listener);
 
-    int getID();
+    long getID();
 
     boolean isDeadlocked();
